requirements.txt
asgiref==3.8.1
crispy-bootstrap4==2024.1
Django==5.1.1
django-appconf==1.0.6
django-ckeditor==6.7.1
django-ckeditor-5==0.2.15
django-crispy-forms==2.3
django-js-asset==2.2.0
django-redis==5.4.0
django-select2==8.2.1
djangorestframework==3.15.2
et_xmlfile==2.0.0
numpy==2.1.2
openpyxl==3.1.5
pandas==2.2.3
pillow==10.4.0
python-dateutil==2.9.0.post0
pytz==2024.2
redis==5.2.0
six==1.16.0
sqlparse==0.5.1
tzdata==2024.2
